Use integer instead of float vectors for rbf

+++ src/Tools/Tool_RBF_DistanceField.cpp
@@ -5,14 +5,14 @@
 Tool_RBF_DistanceField::Tool_RBF_DistanceField() 
 {}
 
-void Tool_RBF_DistanceField::addAnchorPoint(const glm::vec3& pos) {
+void Tool_RBF_DistanceField::addAnchorPoint(const glm::ivec3& pos) {
 	m_anchorPts.push_back(pos);
 	m_valuesAtAnchorPts.conservativeResize(m_valuesAtAnchorPts.size() + 1);
 	m_valuesAtAnchorPts(m_valuesAtAnchorPts.size() - 1) = 0.5f;
 }
 
 void Tool_RBF_DistanceField::evaluateRBFOnWorld(std::function<void(const glm::ivec3 & pos)> whatToDoWithPos){
-	RBF<glm::vec3> rbf = RBF<glm::vec3>(m_anchorPts, m_valuesAtAnchorPts, *m_modulingFunction);
+	RBF<glm::ivec3> rbf = RBF<glm::ivec3>(m_anchorPts, m_valuesAtAnchorPts, *m_modulingFunction);
 	BoundingBox worldBB;
 	for (const glm::ivec3& pos : worldBB) {
 		float d = rbf.eval(pos);
@@ -46,10 +46,10 @@ void Tool_RBF_DistanceField::ImGui_Window() {
 
 	ImGui::Text("Anchor points");
 	for (size_t k = 0; k < m_anchorPts.size(); ++k) {
-		glm::vec3& anchorPt = m_anchorPts[k];
+		glm::ivec3& anchorPt = m_anchorPts[k];
 		float& value = m_valuesAtAnchorPts[k];
 		ImGui::PushID(k);
-		bComputePreview |= ImGui::DragFloat3("Pos", &anchorPt.x);
+		bComputePreview |= ImGui::DragInt3("Pos", &anchorPt.x);
 		bComputePreview |= ImGui::DragFloat("Val", &value);
 		ImGui::PopID();
 		ImGui::Separator();
